+++ CloudStorageService/src/main/java/com/diplom/service/UserService.java
@@ -5,27 +5,22 @@ import com.diplom.model.User;
 import com.diplom.model.dto.UserDTO;
 import com.diplom.repository.UserRepository;
 import com.diplom.request.UserRequest;
-import com.diplom.request.UserUpdateRequest;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
-import org.modelmapper.ModelMapper;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import com.diplom.model.Role;
+
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
 public class UserService {
     private final UserRepository repository;
     private final PasswordEncoder passwordEncoder;
-    private final FileService fileService;
-    private final ModelMapper mapper = new ModelMapper(); // Инициализирован ModelMapper
-    private static final Logger logger = LoggerFactory.getLogger(UserService.class);
 
     private UserDTO convertToUserDto(User user) {
-        return mapper.map(user, UserDTO.class);
+        return new UserDTO(user.getUsername(), user.getEmail());
     }
 
     public User save(User user) {
@@ -38,34 +33,35 @@ public class UserService {
 
     public User findById(Long id) {
         return repository.findById(id)
-                .orElseThrow(() -> new UserNotFoundException("User not found with id: " + id));
+                .orElseThrow(() -> new UserNotFoundException("Пользователь с id : " + id + " не найден."));
     }
 
     public User findByUsername(String name) {
         return repository.findByUsername(name)
-                .orElseThrow(() -> new UserNotFoundException("User not found with name: " + name));
+                .orElseThrow(() -> new UserNotFoundException("Пользователь с именем" + name + " не найден."));
     }
 
     public User findByEmail(String email) {
         return repository.findByEmail(email)
-                .orElseThrow(() -> new UserNotFoundException("User not found with email: " + email));
+                .orElseThrow(() -> new UserNotFoundException("Пользователь с email: " + email + " не найден"));
     }
 
     @Transactional
-    public User updateUserProfile(Long id, UserUpdateRequest userUpdateRequest) {
+    public User updateUserProfile(Long id, UserRequest userRequest) {
         User user = findById(id);
-        mapper.map(userUpdateRequest, user);  // обновление данных пользователя
-        logger.info("User {} is updated", id);
+        user.setUsername(userRequest.getUsername());
+        user.setPassword(userRequest.getPassword());
+        user.setEmail(userRequest.getEmail());
         return repository.save(user);
     }
 
     @Transactional
     public User createUser(UserRequest request) {
         if (repository.findByUsername(request.getUsername()).isPresent()) {
-            throw new IllegalArgumentException("User with that username already exists");
+            throw new IllegalArgumentException("Пользователь уже существует");
         }
         if (repository.findByEmail(request.getEmail()).isPresent()) {
-            throw new IllegalArgumentException("User with that email already exists");
+            throw new IllegalArgumentException("Пользователь уже существует");
         }
 
         User newUser = User.builder()
@@ -75,21 +71,16 @@ public class UserService {
                 .role(Role.ROLE_USER)
                 .build();
 
-        logger.info("New user is created: {}", newUser.getUsername());
         return repository.save(newUser);
     }
 
     @Transactional
-    public void addFileToUsersUploadedFiles(Long fileAuthorId, Long id) {
-        User user = findById(fileAuthorId);
-        user.addFileToUsersUploadedFiles(id);
-        repository.save(user);
-        logger.info("User {} uploaded file {}", fileAuthorId, id);
-    }
-
-    @Transactional
-    public void deleteUser(Long userId) {
-        repository.deleteById(userId);
-        logger.info("User {} is deleted", userId);
+    public void deleteUser (Long userId) {
+        Optional<User> user = repository.findById(userId);
+        if (user.isPresent()) {
+            repository.deleteById(userId);
+        } else {
+            throw new IllegalArgumentException("Пользователь не найден");
+        }
     }
 }
